+++ Protocols/CCoInfyETolerance.cc
@@ -30,7 +30,7 @@ void CCoInfyETolerance::TolerateNode(LinkPtr messageLink)
 
 	if (detector->commonNeighbors.find(infected->id) == detector->commonNeighbors.end())
 		return;
-	map<int, vector<LinkPtr> >::iterator mapIt = detector->commonNeighbors.begin();
+	map<long, vector<LinkPtr> >::iterator mapIt = detector->commonNeighbors.begin();
 	for (; mapIt != detector->commonNeighbors.end(); mapIt++)
 	{
 		if (mapIt->first == infected->id)
